requirements.txt
pandas>=2.0
geopandas>=0.14
shapely>=2.0
pyproj>=3.6
numpy>=1.24
PyYAML>=6.0
